[#34025175] - Unificada versão funcional do código em um único script
